requirements.txt
streamlit
serpapi
python-dotenv
google-generativeai


